Refactor phase change logic.
The goal is to move all "Should a phase change be executed?" logic to the
`next_phase` match expression. This patch does that for just the `add` command.

It may help to understand the `add` command. This command is executed when a
user types `.add` in Discord chat. A user will do this when they want to be
added to the pool of players available for drafting into a team. The phase
where players add themselves is called `PlayerRegistration`. For obvious
reasons, we do not want to move from `PlayerRegistration` to the next
phase (`CaptainSelection`) until the draft pool[1] is full.

So in order to determine whether the next phase should be entered, we want to
check the number of users in the draft pool against the max size of the draft
pool. If the two are equal, then we phase-change.

Prior to this patch, that checking logic existed in the `add` command
itself. This patch moves that logic into the match expression in the
`next_phase` implementation on `Game`. I believe this
approach makes the code easier to reason about due to better separation of
concerns.

[1] Draft pool size is determined by `TEAM_COUNT * TEAM_SIZE`, which are two
environmental variables set in the `.env` file.