Update command! usage to #[command] attribute.
Per the changelog, the command! macro was changes to be an attribute on
a function.

<https://github.com/serenity-rs/serenity/blob/current/CHANGELOG.md#major-breaking-changes>

Looks as though later on the attr macro was updated to require an async
function instead of a blocking one.

Commands now return a Result, but I'm not sure of what happens with the
return values beyond being tested for Ok/Err.

Note that there's a long list of additional attributes you can set when
using the `#[command]` attr:

<https://docs.rs/serenity/0.10.5/serenity/framework/standard/macros/attr.command.html#options>

Likely we'll want to take advantage of these. Help/usage and arg
validation seem especially helpful.