Rework `client_setup()`.
This diff focuses on closing the API gap for serenity for the
initialization of the "framework" code.

Ironically, the previous PR that focused on commands ended up being
insufficient and much of the setup code refactored here ended up being moved
down into the command module(s).

Noted in the code comments for `pugbot::commands`, many of the macros
the serenity framework use rely on name mangling, which sort of begs us
to build up groups, commands, and configure them in the framework within
a single module's scope (so as to avoid having to import the mangled
names in our app code). I don't have a solid recommendation just yet,
but floated some weak ideas in the comments.